+++ ios/versioned/sdk44/EXAdsAdMob/EXAdsAdMob/ABI44_0_0EXAdsAdMobInterstitial.m
@@ -7,22 +7,23 @@ static NSString *const ABI44_0_0EXAdsAdMobInterstitialDidLoad = @"interstitialDi
 static NSString *const ABI44_0_0EXAdsAdMobInterstitialDidFailToLoad = @"interstitialDidFailToLoad";
 static NSString *const ABI44_0_0EXAdsAdMobInterstitialDidOpen = @"interstitialDidOpen";
 static NSString *const ABI44_0_0EXAdsAdMobInterstitialDidClose = @"interstitialDidClose";
-static NSString *const ABI44_0_0EXAdsAdMobInterstitialWillLeaveApplication = @"interstitialWillLeaveApplication";
+static NSString *const ABI44_0_0EXAdsAdMobInterstitialDidFailToOpen = @"interstitialDidFailToOpen";
 
 @interface ABI44_0_0EXAdsAdMobInterstitial ()
 
 @property (nonatomic, weak) id<ABI44_0_0EXEventEmitterService> eventEmitter;
 @property (nonatomic, weak) id<ABI44_0_0EXUtilitiesInterface> utilities;
+@property (nonatomic, strong) GADInterstitialAd *ad;
 
 @end
 
 @implementation ABI44_0_0EXAdsAdMobInterstitial {
-  GADInterstitial  *_interstitial;
   NSString *_adUnitID;
   bool _hasListeners;
   ABI44_0_0EXPromiseResolveBlock _showAdResolver;
   ABI44_0_0EXPromiseResolveBlock _requestAdResolver;
   ABI44_0_0EXPromiseRejectBlock _requestAdRejecter;
+  ABI44_0_0EXPromiseRejectBlock _showAdRejecter;
 }
 
 ABI44_0_0EX_EXPORT_MODULE(ExpoAdsAdMobInterstitialManager);
@@ -40,7 +41,7 @@ ABI44_0_0EX_EXPORT_MODULE(ExpoAdsAdMobInterstitialManager);
            ABI44_0_0EXAdsAdMobInterstitialDidFailToLoad,
            ABI44_0_0EXAdsAdMobInterstitialDidOpen,
            ABI44_0_0EXAdsAdMobInterstitialDidClose,
-           ABI44_0_0EXAdsAdMobInterstitialWillLeaveApplication,
+           ABI44_0_0EXAdsAdMobInterstitialDidClose,
            ];
 }
 
@@ -72,35 +73,77 @@ ABI44_0_0EX_EXPORT_METHOD_AS(requestAd,
                     resolver:(ABI44_0_0EXPromiseResolveBlock)resolve
                     rejecter:(ABI44_0_0EXPromiseRejectBlock)reject)
 {
-  if ([_interstitial hasBeenUsed] || _interstitial == nil) {
-    _requestAdResolver = resolve;
-    _requestAdRejecter = reject;
-    
-    _interstitial = [[GADInterstitial alloc] initWithAdUnitID:_adUnitID];
-    _interstitial.delegate = self;
-    
-    GADRequest *request = [GADRequest request];
-    if (additionalRequestParams) {
-      GADExtras *extras = [[GADExtras alloc] init];
-      extras.additionalParameters = additionalRequestParams;
-      [request registerAdNetworkExtras:extras];
-    }
-    [_interstitial loadRequest:request];
+  ABI44_0_0EX_WEAKIFY(self)
+  if (_ad) {
+    dispatch_async(dispatch_get_main_queue(), ^{
+      ABI44_0_0EX_ENSURE_STRONGIFY(self);
+
+      NSError *error;
+      if ([self.ad canPresentFromRootViewController:self.utilities.currentViewController error:&error]) {
+        [self requestAdWithParams:additionalRequestParams
+                         resolver:resolve
+                         rejecter:reject];
+      } else {
+        reject(@"E_AD_CANNOT_PRESENT", @"Add cannot be presented", error);
+      }
+    });
   } else {
-    reject(@"E_AD_ALREADY_LOADED", @"Ad is already loaded.", nil);
+    [self requestAdWithParams:additionalRequestParams
+                     resolver:resolve
+                     rejecter:reject];
   }
+}
+
+- (void)requestAdWithParams:(NSDictionary *)additionalRequestParams
+                   resolver:(ABI44_0_0EXPromiseResolveBlock)resolve
+                   rejecter:(ABI44_0_0EXPromiseRejectBlock)reject
+{
+  _requestAdResolver = resolve;
+  _requestAdRejecter = reject;
+
+
+  GADRequest *request = [GADRequest request];
+  if (additionalRequestParams) {
+    GADExtras *extras = [[GADExtras alloc] init];
+    extras.additionalParameters = additionalRequestParams;
+    [request registerAdNetworkExtras:extras];
+  }
+
+  ABI44_0_0EX_WEAKIFY(self);
+  [GADInterstitialAd loadWithAdUnitID:_adUnitID
+                              request:request
+                    completionHandler:^(GADInterstitialAd * _Nullable interstitialAd,
+                                        NSError * _Nullable error) {
+    ABI44_0_0EX_ENSURE_STRONGIFY(self);
+
+    if (error) {
+      [self _maybeSendEventWithName:ABI44_0_0EXAdsAdMobInterstitialDidFailToLoad body:@{ @"name": [error description] }];
+      self->_requestAdRejecter(@"E_AD_REQUEST_FAILED", [error description], error);
+      [self cleanupRequestAdPromise];
+      return;
+    }
+
+    self.ad = interstitialAd;
+    self.ad.fullScreenContentDelegate = self;
+
+    [self _maybeSendEventWithName:ABI44_0_0EXAdsAdMobInterstitialDidLoad body:nil];
+    self->_requestAdResolver(nil);
+    [self cleanupRequestAdPromise];
+  }];
+
 }
 
 ABI44_0_0EX_EXPORT_METHOD_AS(showAd,
                     showAd:(ABI44_0_0EXPromiseResolveBlock)resolve
                     rejecter:(ABI44_0_0EXPromiseRejectBlock)reject)
 {
-  if ([_interstitial isReady] && _showAdResolver == nil) {
+  if (_ad && _showAdResolver == nil) {
     _showAdResolver = resolve;
+    _showAdRejecter = reject;
     ABI44_0_0EX_WEAKIFY(self);
     dispatch_async(dispatch_get_main_queue(), ^{
       ABI44_0_0EX_ENSURE_STRONGIFY(self);
-      [self->_interstitial presentFromRootViewController:self.utilities.currentViewController];
+      [self.ad presentFromRootViewController:self.utilities.currentViewController];
     });
   } else if (_showAdResolver != nil) {
     reject(@"E_AD_ALREADY_SHOWING", @"An ad is already being shown, await the first promise.", nil);
@@ -119,9 +162,9 @@ ABI44_0_0EX_EXPORT_METHOD_AS(dismissAd,
     UIViewController *presentedViewController = self.utilities.currentViewController;
     if (presentedViewController != nil && [NSStringFromClass([presentedViewController class]) hasPrefix:@"GAD"]) {
       [presentedViewController dismissViewControllerAnimated:true completion:^{
-        resolve(nil);
         ABI44_0_0EX_ENSURE_STRONGIFY(self);
-        self->_interstitial = nil;
+        self.ad = nil;
+        resolve(nil);
       }];
     } else {
       reject(@"E_AD_NOT_SHOWN", @"Ad is not being shown.", nil);
@@ -133,41 +176,38 @@ ABI44_0_0EX_EXPORT_METHOD_AS(getIsReady,
                     getIsReady:(ABI44_0_0EXPromiseResolveBlock)resolve
                     rejecter:(ABI44_0_0EXPromiseRejectBlock)reject)
 {
-  resolve([NSNumber numberWithBool:[_interstitial isReady]]);
+  resolve([NSNumber numberWithBool:_ad != nil]);
 }
 
-
-- (void)interstitialDidReceiveAd:(GADInterstitial *)ad {
-  [self _maybeSendEventWithName:ABI44_0_0EXAdsAdMobInterstitialDidLoad body:nil];
-  _requestAdResolver(nil);
-  [self _cleanupRequestAdPromise];
-}
-
-- (void)interstitial:(GADInterstitial *)interstitial didFailToReceiveAdWithError:(GADRequestError *)error {
-  [self _maybeSendEventWithName:ABI44_0_0EXAdsAdMobInterstitialDidFailToLoad body:@{ @"name": [error description] }];
-  _requestAdRejecter(@"E_AD_REQUEST_FAILED", [error description], error);
-  [self _cleanupRequestAdPromise];
-  _interstitial = nil;
-}
-
-- (void)interstitialWillPresentScreen:(GADInterstitial *)ad {
-  [self _maybeSendEventWithName:ABI44_0_0EXAdsAdMobInterstitialDidOpen body:nil];
-  _showAdResolver(nil);
-  _showAdResolver = nil;
-}
-
-- (void)interstitialDidDismissScreen:(GADInterstitial *)ad {
-  [self _maybeSendEventWithName:ABI44_0_0EXAdsAdMobInterstitialDidClose body:nil];
-}
-
-- (void)interstitialWillLeaveApplication:(GADInterstitial *)ad {
-  [self _maybeSendEventWithName:ABI44_0_0EXAdsAdMobInterstitialWillLeaveApplication body:nil];
-}
-
-- (void)_cleanupRequestAdPromise
+- (void)cleanupRequestAdPromise
 {
   _requestAdResolver = nil;
   _requestAdRejecter = nil;
+}
+
+- (void)cleanupShowAdPromise
+{
+  _showAdResolver = nil;
+  _showAdRejecter = nil;
+}
+
+#pragma mark - GADFullScreenContentDelegate
+
+- (void)adDidPresentFullScreenContent:(id)ad {
+  [self _maybeSendEventWithName:ABI44_0_0EXAdsAdMobInterstitialDidOpen body:nil];
+  _showAdResolver(nil);
+  [self cleanupShowAdPromise];
+
+}
+
+- (void)ad:(id)ad didFailToPresentFullScreenContentWithError:(NSError *)error {
+  [self _maybeSendEventWithName:ABI44_0_0EXAdsAdMobInterstitialDidFailToOpen body:nil];
+  _showAdRejecter(@"E_AD_SHOW_FAILED", @"Ad failed to present full screen content", error);
+  [self cleanupShowAdPromise];
+}
+
+- (void)adDidDismissFullScreenContent:(id)ad {
+  [self _maybeSendEventWithName:ABI44_0_0EXAdsAdMobInterstitialDidClose body:nil];
 }
 
 @end