+++ packages/expo-ads-admob/ios/EXAdsAdMob/EXAdsGAMBannerView.h
@@ -2,7 +2,7 @@
 #import <ExpoModulesCore/EXModuleRegistry.h>
 #import <GoogleMobileAds/GoogleMobileAds.h>
 
-@interface EXAdsDFPBannerView : UIView <GADBannerViewDelegate, GADAppEventDelegate>
+@interface EXAdsGAMBannerView : UIView <GADBannerViewDelegate, GADAppEventDelegate>
 
 @property (nonatomic, copy) NSString *bannerSize;
 @property (nonatomic, copy) NSString *adUnitID;
@@ -15,7 +15,6 @@
 @property (nonatomic, copy) EXDirectEventBlock onAdViewWillPresentScreen;
 @property (nonatomic, copy) EXDirectEventBlock onAdViewWillDismissScreen;
 @property (nonatomic, copy) EXDirectEventBlock onAdViewDidDismissScreen;
-@property (nonatomic, copy) EXDirectEventBlock onAdViewWillLeaveApplication;
 
 - (GADAdSize)getAdSizeFromString:(NSString *)bannerSize;
 - (void)loadBanner;