+++ packages/expo-face-detector/android/src/main/java/expo/modules/facedetector/ExpoFaceDetector.kt
@@ -7,17 +7,12 @@ import androidx.arch.core.util.Function
 
 import com.google.android.gms.tasks.OnCompleteListener
 import com.google.android.gms.tasks.Task
-import com.google.firebase.ml.vision.common.FirebaseVisionImage
-import com.google.firebase.ml.vision.common.FirebaseVisionImageMetadata
-import com.google.firebase.ml.vision.face.FirebaseVisionFaceDetectorOptions
-import com.google.firebase.ml.vision.face.FirebaseVisionFaceDetector
-import com.google.firebase.ml.vision.face.FirebaseVisionFace
-import com.google.firebase.ml.vision.FirebaseVision
-
-import com.google.firebase.ml.vision.common.FirebaseVisionImageMetadata.ROTATION_0
-import com.google.firebase.ml.vision.common.FirebaseVisionImageMetadata.ROTATION_180
-import com.google.firebase.ml.vision.common.FirebaseVisionImageMetadata.ROTATION_270
-import com.google.firebase.ml.vision.common.FirebaseVisionImageMetadata.ROTATION_90
+import com.google.mlkit.vision.common.InputImage
+import com.google.mlkit.vision.common.InputImage.IMAGE_FORMAT_NV21
+import com.google.mlkit.vision.face.Face
+import com.google.mlkit.vision.face.FaceDetection
+import com.google.mlkit.vision.face.FaceDetector
+import com.google.mlkit.vision.face.FaceDetectorOptions
 
 import expo.modules.facedetector.FaceDetectorUtils.serializeFace
 import expo.modules.facedetector.FaceDetectorUtils.rotateFaceX
@@ -37,7 +32,7 @@ private const val RUN_CLASSIFICATIONS_KEY = "runClassifications"
 private const val TRACKING_KEY = "tracking"
 
 class ExpoFaceDetector(private val context: Context) : FaceDetectorInterface {
-  private var faceDetector: FirebaseVisionFaceDetector? = null
+  private var faceDetector: FaceDetector? = null
   private val minFaceSize = 0.15f
   private var minDetectionInterval: Long = 0
   private var lastDetectionMillis: Long = 0
@@ -76,8 +71,8 @@ class ExpoFaceDetector(private val context: Context) : FaceDetectorInterface {
     if (faceDetector == null) {
       createFaceDetector()
     }
-    val image = FirebaseVisionImage.fromFilePath(context, filePath)
-    faceDetector?.detectInImage(image)
+    val image = InputImage.fromFilePath(context, filePath)
+    faceDetector?.process(image)
       ?.addOnCompleteListener(
         faceDetectionHandler(FaceDetectorUtils::serializeFace, complete, error)
       )
@@ -98,23 +93,16 @@ class ExpoFaceDetector(private val context: Context) : FaceDetectorInterface {
     if (faceDetector == null) {
       createFaceDetector()
     }
-    val firRotation = getFirRotation(rotation)
-    val metadata = FirebaseVisionImageMetadata.Builder()
-      .setWidth(width)
-      .setHeight(height)
-      .setFormat(FirebaseVisionImageMetadata.IMAGE_FORMAT_NV21)
-      .setRotation(firRotation)
-      .build()
-    val image = FirebaseVisionImage.fromByteArray(imageData, metadata)
+    val image = InputImage.fromByteArray(imageData, width, height, rotation, IMAGE_FORMAT_NV21)
     if (minDetectionInterval <= 0 || minIntervalPassed()) {
       lastDetectionMillis = System.currentTimeMillis()
-      faceDetector?.detectInImage(image)
+      faceDetector?.process(image)
         ?.addOnCompleteListener(
           faceDetectionHandler(
             { face ->
               var result = serializeFace(face, scaleX, scaleY)
               if (mirrored) {
-                result = if (firRotation == ROTATION_270 || firRotation == ROTATION_90) {
+                result = if (rotation == 270 || rotation == 90) {
                   rotateFaceX(result, height, scaleX)
                 } else {
                   rotateFaceX(result, width, scaleX)
@@ -130,15 +118,8 @@ class ExpoFaceDetector(private val context: Context) : FaceDetectorInterface {
     }
   }
 
-  private fun getFirRotation(rotation: Int) = when ((rotation + 360) % 360) {
-    90 -> ROTATION_90
-    180 -> ROTATION_180
-    270 -> ROTATION_270
-    else -> ROTATION_0
-  }
-
-  private fun faceDetectionHandler(transformer: Function<FirebaseVisionFace, Bundle>, complete: FacesDetectionCompleted, error: FaceDetectionError): OnCompleteListener<List<FirebaseVisionFace>?> =
-    OnCompleteListener { task: Task<List<FirebaseVisionFace>?> ->
+  private fun faceDetectionHandler(transformer: Function<Face, Bundle>, complete: FacesDetectionCompleted, error: FaceDetectionError): OnCompleteListener<List<Face>?> =
+    OnCompleteListener { task: Task<List<Face>?> ->
       if (task.isComplete && task.isSuccessful) {
         val facesArray = ArrayList<Bundle>().apply {
           val faces = task.result
@@ -182,11 +163,11 @@ class ExpoFaceDetector(private val context: Context) : FaceDetectorInterface {
   }
 
   private fun createFaceDetector() {
-    faceDetector = FirebaseVision.getInstance().getVisionFaceDetector(createOptions())
+    faceDetector = FaceDetection.getClient(createOptions())
   }
 
-  private fun createOptions(): FirebaseVisionFaceDetectorOptions {
-    val builder = FirebaseVisionFaceDetectorOptions.Builder()
+  private fun createOptions(): FaceDetectorOptions {
+    val builder = FaceDetectorOptions.Builder()
       .setClassificationMode(classificationType)
       .setLandmarkMode(landmarkType)
       .setPerformanceMode(mode)